requirements.txt
llama-index==0.11.20
llama-index-llms-groq==0.2.0
PyYAML==6.0.2
plotly==5.24.1
streamlit==1.39.0
docx2txt==0.8
llama-index-embeddings-huggingface==0.3.1
transformers==4.46.0
torch==2.5.0
sentence-transformers==3.2.1
fastapi==0.115.6
uvicorn==0.34.0
